fix error show order

+++ src/main/java/com/example/phone_store/mapper/OrderMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Mapper
 public interface OrderMapper {
     void save(Order order);
-    List<Order> findByUserId(@Param("userId") Integer userId);
+    List<Order> findByUserId(@Param("userId") Long userId);
     List<Order> findAll();
     Order findById(@Param("orderId") Integer orderId);
     int updateStatus(@Param("orderId") Integer orderId, @Param("status") String status);
